+++ src/components/AddFavouriteBtn/AddFavouriteBtn.tsx
@@ -1,6 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import classNames from 'classnames';
-import { ReactComponent as HeartIcon } from '../../icons/favourites-icon.svg';
+import {
+  ReactComponent as HeartIcon,
+} from '../../assets/icons/favourites-icon.svg';
 import { ProductItem } from '../../types/ProductItem';
 
 import './addFavouriteBtn.scss';
