+++ src/components/AddFavouriteBtn/addFavouriteBtn.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .add-favourite-btn {
   width: 40px;
