+++ src/components/AddToCartBtn/addToCartBtn.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .add-to-cart {
   background: $gray-primary;
