+++ src/components/BackBtn/backBtn.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .back-btn {
   display: flex;
