+++ src/components/BackToTop/backToTop.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .back-to-top {
   display: flex;
