+++ src/components/Banner/banner.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .banner {
   margin-bottom: 82px;
