+++ src/components/CartItem/cartItem.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .cart-item {
   border: 1px solid $gray-element;
