+++ src/components/Footer/footer.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .footer {
   padding: 42px 0;
