+++ src/components/Pagination/Pagination.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .pagination {
   display: flex;
