+++ src/components/ProductCard/productCard.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .product-card {
   padding: 24px;
