+++ src/components/ProductsList/productList.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .products-list {
   display: grid;
