+++ src/components/ProductsSlider/PrevArrow.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ReactComponent as Arrow } from '../../icons/small-arrow.svg';
+import { ReactComponent as Arrow } from '../../assets/icons/small-arrow.svg';
 
 export const PrevArrow: React.FC = (props: any) => {
   const { onClick } = props;
