+++ src/components/Search/search.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .search {
   border-left: 1px solid $gray-element;
