+++ src/components/ShopCategory/shopCategory.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .shopCategory {
   margin-top: 80px;
