+++ src/pages/Favourites/favourites.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .favourites {
   &__title {
