+++ src/pages/ProductDetails/productDetails.scss
@@ -1,4 +1,4 @@
-@import "../../variables/variables.scss";
+@import "../../assets/styles/variables/variables.scss";
 
 .gallary {
   display: flex;
